+++ lib/src/metrics/export/otlp/otlp_grpc_metric_exporter.dart
@@ -178,6 +178,7 @@ class OtlpGrpcMetricExporter implements MetricExporter {
       MetricTransformer.transformMetrics(
         data,
         fallbackResource: OTel.resource(null),
+        exemplarFilter: config.exemplarFilter,
       );
 
   @override